+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [unreleased]
 
-### Added
+### Custom Leaderboards
 
 - Added 'date created' and 'date last played' columns to custom leaderboards list.
+- UI improvements.
 
-### Changed
+### Miscellaneous
 
 - Updated ImGui.NET dependency from 1.90.6.1 to 1.91.0.1.
 - Minor UI improvements.